combines server files removes unwanted lines add new command in package.json
