+++ packages/client/.gitignore
@@ -11,6 +11,7 @@ lerna-debug.log*
 node_modules
 dist
 dist-ssr
+ssr-dist
 *.local
 
 # Editor directories and files
