+++ packages/client/src/store/User/auth/actions.ts
@@ -6,7 +6,6 @@ import { IUser } from '../../types';
 import { API_ENDPOINTS } from '../../constatns';
 import { errorMessage } from '../../../utils/apiErrorMessageCheck';
 
-// const AUTH_ENDPOINT = `${API_MAIN_ENDPOINT}/auth`;
 interface IUserService {
   getCurrentUser(): Promise<IUser>;
 }
@@ -15,10 +14,8 @@ export const fetchAuthUserGet = createAsyncThunk(
   'auth/fetchAuth',
   async (_, thunkAPI) => {
     try {
-      // const response = await axios.get<IUser>(`${API_ENDPOINTS.auth}/user`);
-
-      // return response.data;
       const service: IUserService = thunkAPI.extra as IUserService;
+
       return await service.getCurrentUser();
     } catch (error) {
       return thunkAPI.rejectWithValue(
