+++ packages/server/app/app.ts
@@ -1,0 +1,12 @@
+import express from 'express';
+import { proxy } from '../middlewares/proxy';
+import { createClientAndConnect } from '../db';
+import cors from 'cors';
+
+const app = express();
+app.use(cors());
+app.use('/api/v2/*', proxy);
+
+createClientAndConnect();
+
+export default app;
