+++ packages/server/middlewares/proxy.ts
@@ -1,0 +1,9 @@
+import { createProxyMiddleware } from 'http-proxy-middleware';
+
+export const proxy = createProxyMiddleware({
+  changeOrigin: true,
+  cookieDomainRewrite: {
+    '*': '',
+  },
+  target: 'https://ya-praktikum.tech',
+});
