+++ packages/server/repository/YandexAPIRepository.ts
@@ -6,15 +6,12 @@ export class YandexAPIRepository {
   constructor(private _cookieHeader: string | undefined) {}
 
   async getCurrent(): Promise<any> {
-    console.log('Запрос с сервера');
+    console.log('Запрос пользователя с сервера');
     const { data } = await axios.get(`${API_ROOT}/auth/user`, {
       headers: {
         cookie: this._cookieHeader,
       },
     });
-    return {
-      ...data,
-      // xss: "</script><script>alert('pwned')</script><!--",
-    };
+    return data;
   }
 }
